(Dom) Energy Assault finished, adjusted Energy Focus inherent
I gave an effect to the Inherent.Energy_Focus, which when toggled on will add a 100% chance to the GlobalChanceMod called Energy (if we can add an enum of EnergyFocus or a GlobalChanceMod called EnergyFocus, we'll adjust this change).
I made all Energy Focus required effects have 0% chance and tagged it with GlobalChanceMod of "Energy". Works like a charm